Ordenamos estructura
Antes de continuar escalando la funcionalidad ordenamos la estructura de
paquetes como corresponde para desarrollar codigo mas limpio y ordenado.
