+++ src/main/java/com/capitole/prices/application/dto/PriceByCriteriaRequest.java
@@ -1,10 +1,9 @@
-package com.capitole.prices.application;
+package com.capitole.prices.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 import java.time.LocalDateTime;
 
 @Data
