+++ src/main/java/com/capitole/prices/application/service/PriceServiceImpl.java
@@ -1,27 +1,28 @@
-package com.capitole.prices.application;
+package com.capitole.prices.application.service;
 
 import java.util.List;
-import java.util.Optional;
 
-import com.capitole.prices.infrastructure.adapters.repositories.PriceRepository;
+import com.capitole.prices.application.dto.PriceByCriteriaRequest;
+import com.capitole.prices.application.dto.PriceByCriteriaResponse;
+import com.capitole.prices.infrastructure.adapters.repositories.H2PriceRepository;
 import org.springframework.stereotype.Service;
 import com.capitole.prices.domain.Price;
-import com.capitole.prices.domain.PriceService;
+import com.capitole.prices.domain.services.PriceService;
 import com.capitole.prices.infrastructure.exceptions.PriceNotFoundException;
 
 @Service
 public class PriceServiceImpl implements PriceService {
 
-	private final PriceRepository priceRepository;
+	private final H2PriceRepository priceRepository;
 
-	public PriceServiceImpl(PriceRepository priceRepository) {
+	public PriceServiceImpl(H2PriceRepository priceRepository) {
 		this.priceRepository = priceRepository;
 	}
 
 
 	@Override
 	public PriceByCriteriaResponse findByCriteria(PriceByCriteriaRequest priceByCriteriaRequest) {
-		List<Price> optionalPrice = this.priceRepository.findTopByBrandIdAndProductIdAndDateOrderByPriorityDesc(
+		List<Price> optionalPrice = this.priceRepository.findByBrandIdAndProductIdAndDateOrderByPriorityDesc(
 				priceByCriteriaRequest.getBrandIdentifier(),
 				priceByCriteriaRequest.getProductIdentifier(),
 				priceByCriteriaRequest.getApplicationDate());
