--- src/main/java/com/capitole/prices/domain/PriceService.java
@@ -1,11 +1,0 @@
-package com.capitole.prices.domain;
-
-
-import com.capitole.prices.application.PriceByCriteriaRequest;
-import com.capitole.prices.application.PriceByCriteriaResponse;
-
-public interface PriceService {
-	
-	public PriceByCriteriaResponse findByCriteria(PriceByCriteriaRequest priceByCriteriaRequest);
-
-}
